requirements.txt
streamlit>=1.32
openai>=1.25
pillow>=10.0
pdf2image>=1.17
pymupdf